split out codegen logic
