+++ .circleci/config.yml
@@ -21,7 +21,7 @@ jobs:
       - checkout
       - restore_cache:
           key: sbtcache
-      - run: sbt ++2.12.10! coreJVM/test http4s/compile akkaHttp/compile finch/compile uzhttp/compile examples/compile catsInteropJVM/compile benchmarks/compile codegen/test clientJVM/test monixInterop/compile federation/test codegenSbt/test
+      - run: sbt ++2.12.10! coreJVM/test http4s/compile akkaHttp/compile finch/compile uzhttp/compile examples/compile catsInteropJVM/compile benchmarks/compile codegen/test codegenStb/test clientJVM/test monixInterop/compile federation/test
       - save_cache:
           key: sbtcache
           paths:
